Fixed a typo.

+++ src/runtime/mod.rs
@@ -19,7 +19,7 @@ pub enum Value {
     I32(i32),
     I64(i64),
     F32(f32),
-    F64(f32),
+    F64(f64),
     BlockLock,
 }
